Update to remove warnings, fix new 8266 core rev
Remove build warnings and use the faster pgm_read_*_aligned macros to
update to the latest ESP8266 Arduino core.

+++ src/AudioFileSourceBuffer.cpp
@@ -162,7 +162,7 @@ void AudioFileSourceBuffer::fill()
       int cnt = src->readNonBlock(&buffer[writePtr], bytesAvailEnd);
       length += cnt;
       writePtr = (writePtr + cnt) % buffSize;
-      if (cnt != bytesAvailEnd) return;
+      if (cnt != (int)bytesAvailEnd) return;
     }
 
     if (readPtr > 1) {
